+++ packages/transform/src/features/llm/LlmApplicationTransformer.ts
@@ -5,7 +5,7 @@ import {
   MetadataFactory,
   MetadataSchema,
 } from "@typia/core";
-import { ILlmApplication, ILlmSchema, ValidationPipe } from "@typia/interface";
+import { ILlmSchema, ValidationPipe } from "@typia/interface";
 import ts from "typescript";
 
 import { ITransformProps } from "../../ITransformProps";
@@ -86,93 +86,4 @@ export namespace LlmApplicationTransformer {
     });
   };
 
-  /** @internal */
-  export const decompose = (
-    method: string,
-    props: ITransformProps,
-  ): {
-    application: ILlmApplication.__IPrimitive;
-    type: ts.Type;
-    node: ts.TypeNode;
-    config:
-      | Partial<
-          ILlmSchema.IConfig & {
-            equals: boolean;
-          }
-        >
-      | undefined;
-  } | null => {
-    // GET GENERIC ARGUMENT
-    if (!props.expression.typeArguments?.length)
-      throw new TransformerError({
-        code: `typia.llm.${method}`,
-        message: "no generic argument.",
-      });
-    const top: ts.Node = props.expression.typeArguments[0]!;
-    if (ts.isTypeNode(top) === false) return null;
-
-    // GET TYPE
-    const config:
-      | Partial<
-          ILlmSchema.IConfig & {
-            equals: boolean;
-          }
-        >
-      | undefined = LlmMetadataFactory.getConfig({
-      context: props.context,
-      method,
-      node: props.expression.typeArguments[1],
-    });
-    const type: ts.Type = props.context.checker.getTypeFromTypeNode(top);
-
-    // VALIDATE TYPE
-    const analyze = (validate: boolean): MetadataSchema => {
-      const result: ValidationPipe<MetadataSchema, MetadataFactory.IError> =
-        MetadataFactory.analyze({
-          checker: props.context.checker,
-          transformer: props.context.transformer,
-          options: {
-            absorb: validate,
-            escape: true,
-            constant: true,
-            functional: true,
-            validate:
-              validate === true
-                ? (next) =>
-                    LlmApplicationProgrammer.validate({
-                      config,
-                      metadata: next.metadata,
-                      explore: next.explore,
-                      top: next.top,
-                    })
-                : undefined,
-          },
-          components: new MetadataCollection({
-            replace: MetadataCollection.replace,
-          }),
-          type,
-        });
-      if (result.success === false)
-        throw TransformerError.from({
-          code: `typia.llm.${method}`,
-          errors: result.errors,
-        });
-      return result.data;
-    };
-    analyze(true);
-
-    // GENERATE LLM APPLICATION
-    return {
-      application: LlmApplicationProgrammer.writeApplication({
-        context: props.context,
-        modulo: props.modulo,
-        metadata: analyze(false),
-        config,
-        name: top.getFullText().trim(),
-      }),
-      node: top,
-      type,
-      config,
-    };
-  };
 }
