+++ packages/transform/src/features/llm/LlmControllerTransformer.ts
@@ -1,75 +1,100 @@
-import { LiteralFactory } from "@typia/core";
+import {
+  LlmApplicationProgrammer,
+  LlmControllerProgrammer,
+  LlmMetadataFactory,
+  MetadataCollection,
+  MetadataFactory,
+  MetadataSchema,
+} from "@typia/core";
+import { ILlmSchema, ValidationPipe } from "@typia/interface";
 import ts from "typescript";
 
 import { ITransformProps } from "../../ITransformProps";
 import { TransformerError } from "../../TransformerError";
-import { LlmApplicationTransformer } from "./LlmApplicationTransformer";
 
 export namespace LlmControllerTransformer {
   export const transform = (props: ITransformProps): ts.Expression => {
-    const dec = LlmApplicationTransformer.decompose("application", props);
-    if (dec === null) return props.expression;
-    else if (props.expression.arguments[0] === undefined)
+    // GET GENERIC ARGUMENT
+    if (!props.expression.typeArguments?.length)
       throw new TransformerError({
-        code: `typia.llm.controller`,
-        message: `no identifier name.`,
+        code: "typia.llm.controller",
+        message: "no generic argument.",
       });
-    else if (props.expression.arguments[1] === undefined)
+    const top: ts.Node = props.expression.typeArguments[0]!;
+    if (ts.isTypeNode(top) === false) return props.expression;
+
+    if (props.expression.arguments[0] === undefined)
       throw new TransformerError({
-        code: `typia.llm.controller`,
-        message: `no executor.`,
+        code: "typia.llm.controller",
+        message: "no identifier name.",
+      });
+    if (props.expression.arguments[1] === undefined)
+      throw new TransformerError({
+        code: "typia.llm.controller",
+        message: "no executor.",
       });
 
-    const typeNode: ts.ImportTypeNode = props.context.importer.type({
-      file: "typia",
-      name: "ILlmController",
-      arguments: [dec.node],
+    // GET CONFIG
+    const config:
+      | Partial<
+          ILlmSchema.IConfig & {
+            equals: boolean;
+          }
+        >
+      | undefined = LlmMetadataFactory.getConfig({
+      context: props.context,
+      method: "controller",
+      node: props.expression.typeArguments[1],
     });
-    const primitiveTypeNode: ts.ImportTypeNode = props.context.importer.type({
-      file: "typia",
-      name: "ILlmApplication.__IPrimitive",
-      arguments: [dec.node],
+    const type: ts.Type = props.context.checker.getTypeFromTypeNode(top);
+
+    // VALIDATE TYPE
+    const analyze = (validate: boolean): MetadataSchema => {
+      const result: ValidationPipe<MetadataSchema, MetadataFactory.IError> =
+        MetadataFactory.analyze({
+          checker: props.context.checker,
+          transformer: props.context.transformer,
+          options: {
+            absorb: validate,
+            escape: true,
+            constant: true,
+            functional: true,
+            validate:
+              validate === true
+                ? (next) =>
+                    LlmApplicationProgrammer.validate({
+                      config,
+                      metadata: next.metadata,
+                      explore: next.explore,
+                      top: next.top,
+                    })
+                : undefined,
+          },
+          components: new MetadataCollection({
+            replace: MetadataCollection.replace,
+          }),
+          type,
+        });
+      if (result.success === false)
+        throw TransformerError.from({
+          code: "typia.llm.controller",
+          errors: result.errors,
+        });
+      return result.data;
+    };
+    analyze(true);
+
+    // GENERATE LLM CONTROLLER
+    return LlmControllerProgrammer.write({
+      context: props.context,
+      modulo: props.modulo,
+      metadata: analyze(false),
+      className: top.getFullText().trim(),
+      config,
+      node: top,
+      nameArgument: props.expression.arguments[0],
+      executeArgument: props.expression.arguments[1],
+      configArgument: props.expression.arguments[2],
     });
-    const value: ts.ObjectLiteralExpression =
-      ts.factory.createObjectLiteralExpression(
-        [
-          ts.factory.createPropertyAssignment(
-            "protocol",
-            ts.factory.createStringLiteral("class"),
-          ),
-          ts.factory.createPropertyAssignment(
-            "name",
-            props.expression.arguments[0],
-          ),
-          ts.factory.createPropertyAssignment(
-            "execute",
-            props.expression.arguments[1],
-          ),
-          ts.factory.createPropertyAssignment(
-            "application",
-            ts.factory.createCallExpression(
-              props.context.importer.internal("llmApplicationFinalize"),
-              undefined,
-              [
-                ts.factory.createAsExpression(
-                  ts.factory.createSatisfiesExpression(
-                    LiteralFactory.write(dec.application),
-                    primitiveTypeNode,
-                  ),
-                  primitiveTypeNode,
-                ),
-                ...(props.expression.arguments?.[2] !== undefined
-                  ? [props.expression.arguments?.[2]]
-                  : []),
-              ],
-            ),
-          ),
-        ],
-        true,
-      );
-    return ts.factory.createAsExpression(
-      ts.factory.createSatisfiesExpression(value, typeNode),
-      typeNode,
-    );
   };
 }
